[REFACTOR][E] Move context classes to context package
Creates the context package and moves the Eclipse plugin and session
context extensions to the package.

Adds a package-info for the package.
