+++ eclipse/src/saros/context/package-info.java
@@ -1,0 +1,2 @@
+/** This package contains the Saros context extensions for the Eclipse implementation. */
+package saros.context;
