--- eclipse/src/saros/project/internal/package-info.java
@@ -1,5 +1,0 @@
-/**
- * This package offers implementations of several classes in the parent package, in addition to
- * managers for things such as colors, permissions, newly added projects.
- */
-package saros.project.internal;
