+++ eclipse/test/junit/saros/SarosEclipseContextFactoryTest.java
@@ -10,6 +10,7 @@ import org.powermock.core.classloader.annotations.PrepareForTest;
 import org.powermock.modules.junit4.PowerMockRunner;
 import saros.context.CoreContextFactory;
 import saros.context.IContextFactory;
+import saros.context.SarosEclipseContextFactory;
 import saros.repackaged.picocontainer.MutablePicoContainer;
 import saros.test.mocks.ContextMocker;
 import saros.test.mocks.EclipseMocker;
