+++ eclipse/test/junit/saros/session/internal/SarosSessionTest.java
@@ -40,6 +40,7 @@ import saros.awareness.AwarenessInformationCollector;
 import saros.context.CoreContextFactory;
 import saros.context.IContainerContext;
 import saros.context.IContextKeyBindings;
+import saros.context.SarosEclipseSessionContextFactory;
 import saros.editor.EditorManager;
 import saros.filesystem.IPathFactory;
 import saros.net.IConnectionManager;
@@ -50,7 +51,6 @@ import saros.net.xmpp.JID;
 import saros.net.xmpp.XMPPConnectionService;
 import saros.preferences.EclipsePreferences;
 import saros.preferences.PreferenceStore;
-import saros.project.internal.SarosEclipseSessionContextFactory;
 import saros.repackaged.picocontainer.BindKey;
 import saros.repackaged.picocontainer.MutablePicoContainer;
 import saros.repackaged.picocontainer.PicoBuilder;
